+++ spec/integration/rails/rails_spec.rb
@@ -301,7 +301,9 @@ RSpec.describe "Rails integration specs" do
 
       wait_for_a_request_with_body(/"errors"/)
 
-      body = %r|{"routes":\[{"method":"GET","route":"\/crash\(\.:format\)"|
+      body = %r|
+        {"routes":\[{"method":"GET","route":"\/crash\(\.:format\)","status_code":500
+      |x
       wait_for(a_request(:put, routes_endpoint).with(body: body)).
         to have_been_made.once
     end